creates playlist in spotify through create playlist form
