+++ screens/TestRoom.js
@@ -32,54 +32,12 @@ export default function TestRoom(props) {
 //       .then(result => setRoom(result))
 //       .catch(err => console.log(err))
 //   }, []);
-  const handleSubmit = async() => {
-    try{
-        let result = await enterRoom(authData.passcode, authData.title, authData.userName)
-        if(result === 'Invalid credentials'){
-            Alert.alert(
-                'Alert Title',
-                result + ' try again',
-                {cancelable: false},
-              );
-        }else {
-            props.navigation.navigate('PlaylistRoom');
-        }
-    }catch(err){
-        console.log(err);
-    }
-  }
   return (
     <ScrollView>
         <View style={styles.conatiner}>
         <View>
-          <Text style={styles.header}>Join A Playlist</Text>
+          <Text style={styles.header}>Add Freebird to your playlist</Text>
         </View>
-          <View style={styles.inputContainer}>
-            <TextInput
-              style={styles.textInput}
-              placeholder="Playlist Title"
-              maxLength={100}
-              onChangeText = {text => setAuthData({...authData, title: text})}
-            />
-            <TextInput
-              style={styles.textInput}
-              placeholder="Your Name"
-              maxLength={50}
-              onChangeText = {text => setAuthData({...authData, userName: text})}
-            />
-            <TextInput
-              style={styles.textInput}
-              placeholder="Playlist Passcode"
-              maxLength={4}
-              onChangeText = {text => setAuthData({...authData, passcode: text})}
-            />
-          </View>
-          <TouchableOpacity
-            style={styles.saveButton}
-            onPress={() => handleSubmit()}
-          >
-            <Text style={styles.saveButtonText}>Enter</Text>
-          </TouchableOpacity>
       </View>
     </ScrollView>
   );
